perf: optimize polynomial division
Computing a the inverse in `Fp` is the most expensive step in `Fp`
division, which in turn takes most of the time in the polynomial
division itself.
Extracting this out of the hot loop and replacing the call site with a
product gives a huge boost (70-90% measured in polynomial benchmarks and
almost 10% in STARK proving benchmarks).