fix: substraction overflow in `sqrt`
An edge case where the inner loop exits without finding a power such
that it becomes `one` makes `i == m` which in turn causes a later
`m - i - 1` to overflow.
This was caused by the demotion to primitive type in a previous
optimization.
The current fix is reverting those to become `FieldElement`s again.